fix: int型への変換と、リダイレクト先の統一化

+++ health/delete.php
@@ -15,7 +15,7 @@ if ($_SERVER['REQUEST_METHOD'] !== 'POST') {
 }
 
 // POSTリクエストからIDを取得
-$id = $_POST['id'] ?? 0;
+$id = (int) ($_POST['id'] ?? 0);
 
 if ($id > 0) {
     // データベース接続
